save owner user to project

+++ src/controllers/ProjectController.ts
@@ -1,9 +1,11 @@
-import type { Request, Response } from "express";
-import Project from "../models/Project";
+import type { Request, Response } from "express"
+import Project from "../models/Project"
 
 export class ProjectController {
   static createProject = async (req: Request, res: Response) => {
     const project = new Project(req.body);
+
+    project.manager = req.user.id;
 
     try {
       await project.save();
@@ -15,7 +17,9 @@ export class ProjectController {
 
   static getAllProjects = async (req: Request, res: Response) => {
     try {
-      const projects = await Project.find({});
+      const projects = await Project.find({
+        $or: [{ manager: { $in: req.user.id } }],
+      });
       res.json(projects);
     } catch (error) {
       console.error(error);
@@ -28,6 +32,10 @@ export class ProjectController {
       const project = await Project.findById(id).populate("tasks");
 
       if (!project) return res.status(404).json({ error: "No project found" });
+
+      if (project.manager.toString() !== req.user.id.toString()) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
 
       res.json(project);
     } catch (error) {
@@ -42,6 +50,10 @@ export class ProjectController {
 
       if (!project) return res.status(404).json({ error: "No project found" });
 
+      if (project.manager.toString() !== req.user.id.toString()) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
+
       await project.save();
 
       res.json(project);
@@ -55,6 +67,9 @@ export class ProjectController {
     try {
       const project = await Project.findById(id);
       if (!project) return res.status(404).json({ error: "No project found" });
+      if (project.manager.toString() !== req.user.id.toString()) {
+        return res.status(401).json({ error: "Unauthorized" });
+      }
       await project.deleteOne();
       res.send("Project deleted");
     } catch (error) {
